move css to style
