+++ src/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import './style/App.css';
 
 function App() {
   return (
